requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
cssselect==1.0.3
Django==2.2.3
django-haystack==2.8.1
django-redis==4.10.0
django-tinymce==2.7.0
dnspython==1.15.0
easygui==0.96
eventlet==0.24.1
greenlet==0.4.15
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
itsdangerous==1.1.0
jieba==0.39
kombu==4.2.1
lxml==4.3.4
Markdown==3.1.1
monotonic==1.5
mutagen==1.41.1
mysqlclient==1.4.2.post1
parsel==1.5.1
Pillow==5.3.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycryptodomex==3.7.2
PyDispatcher==2.0.5
pygame==1.9.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyMySQL==0.9.2
pyOpenSSL==19.0.0
pypiwin32==223
python-alipay-sdk==1.8.0
pytz==2018.7
pywin32==224
queuelib==1.5.0
redis==2.10.6
requests==2.20.1
reverse==0.1.0
Scrapy==1.7.2
service-identity==18.1.0
six==1.11.0
sqlparse==0.3.0
Twisted==19.2.0
urllib3==1.24.1
vine==1.1.4
w3lib==1.20.0
Whoosh==2.7.4
zope.interface==4.6.0
